requirements.txt
runpod
# torch/torchvision installed in Dockerfile (nightly cu124 for sm_120); do not reinstall here
# torch
# torchvision
# torchaudio
accelerate>=0.26.0
qwen-vl-utils
pillow
requests
pdf2image
git+https://github.com/huggingface/transformers.git
transformers-stream-generator
einops
tiktoken
# flash-attn --no-build-isolation
